backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic
pydantic-settings
python-dotenv
python-multipart
python-jose[cryptography]
passlib[bcrypt]
bcrypt==3.1.7
cors
openai>=1.0.0
alembic
google-auth
pydantic[email]
livekit-api
livekit-agents[google]==1.3.12
livekit-plugins-openai
livekit-plugins-google==1.3.12
livekit-plugins-silero